Added optional parameter to provide assembly for version checking.
Added an optional parameter to Start method, this allows you to pass in the Assembly you'd like to use for version checking. The original was failing when the assembly that needed to be checked was called from a main.exe via a dynamic plugin.
Specifically, a Microsoft Office VSTO plugin that I wanted to auto update, when using the original code the GetEntryAssembly() call failed and returned null.